+++ Src/CSharpier.VSCode/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 2.0.0
-- Adding support for CSharpier 1.0.0
+- Adding support for CSharpier 1.0.0 which includes the ability to format xml files.
 
 ## 1.9.2
 - Add option for diagnostic level, default to warning